Moved LoggingConfigurationElementExtensions into its own file
