+++ src/NLog/Config/LoggingConfigurationParser.cs
@@ -125,9 +125,7 @@ namespace NLog.Config
                         LogFactory.ThrowExceptions = ParseBooleanValue(configItem.Key, configItem.Value, LogFactory.ThrowExceptions);
                         break;
                     case "THROWCONFIGEXCEPTIONS":
-                        LogFactory.ThrowConfigExceptions = StringHelpers.IsNullOrWhiteSpace(configItem.Value)
-                            ? (bool?)null
-                            : ParseBooleanValue(configItem.Key, configItem.Value, false);
+                        LogFactory.ThrowConfigExceptions = ParseNullableBooleanValue(configItem.Key, configItem.Value, false);
                         break;
                     case "INTERNALLOGLEVEL":
                         InternalLogger.LogLevel = ParseLogLevelSafe(configItem.Key, configItem.Value, InternalLogger.LogLevel);
@@ -171,14 +169,10 @@ namespace NLog.Config
                         LogFactory.GlobalThreshold = ParseLogLevelSafe(configItem.Key, configItem.Value, LogFactory.GlobalThreshold);
                         break; // expanding variables not possible here, they are created later
                     case "PARSEMESSAGETEMPLATES":
-                        parseMessageTemplates = string.IsNullOrEmpty(configItem.Value)
-                            ? (bool?)null
-                            : ParseBooleanValue(configItem.Key, configItem.Value, true);
+                        parseMessageTemplates = ParseNullableBooleanValue(configItem.Key, configItem.Value, true);
                         break;
                     case "AUTOSHUTDOWN":
-                        LogFactory.AutoShutdown = StringHelpers.IsNullOrWhiteSpace(configItem.Value)
-                                ? true
-                                : ParseBooleanValue(configItem.Key, configItem.Value, true);
+                        LogFactory.AutoShutdown = ParseBooleanValue(configItem.Key, configItem.Value, true);
                         break;
                     case "AUTORELOAD":
                         break;  // Ignore here, used by other logic
@@ -634,14 +628,13 @@ namespace NLog.Config
 
             var rule = new LoggingRule(ruleName)
             {
-                LoggerNamePattern = namePattern
+                LoggerNamePattern = namePattern,
+                Final = final,
             };
 
             EnableLevelsForRule(rule, enableLevels, minLevel, maxLevel);
 
             ParseLoggingRuleTargets(writeTargets, rule);
-
-            rule.Final = final;
 
             ParseLoggingRuleChildren(loggerElement, rule);
 
@@ -795,7 +788,7 @@ namespace NLog.Config
 
             foreach (var targetElement in targetsElement.Children)
             {
-                string targetTypeName = GetConfigItemTypeAttribute(targetElement);
+                string targetTypeName = targetElement.GetConfigItemTypeAttribute();
                 string targetValueName = targetElement.GetOptionalValue("name", null);
                 Target newTarget = null;
                 if (!string.IsNullOrEmpty(targetValueName))
@@ -892,7 +885,7 @@ namespace NLog.Config
         private void ParseTargetElement(Target target, ILoggingConfigurationElement targetElement,
             Dictionary<string, ILoggingConfigurationElement> typeNameToDefaultTargetParameters = null)
         {
-            string targetTypeName = GetConfigItemTypeAttribute(targetElement, "targets");
+            string targetTypeName = targetElement.GetConfigItemTypeAttribute("targets");
             ILoggingConfigurationElement defaults;
             if (typeNameToDefaultTargetParameters != null &&
                 typeNameToDefaultTargetParameters.TryGetValue(targetTypeName, out defaults))
@@ -948,7 +941,7 @@ namespace NLog.Config
 
             if (IsTargetElement(name))
             {
-                string targetTypeName = GetConfigItemTypeAttribute(childElement, GetName(wrapper));
+                string targetTypeName = childElement.GetConfigItemTypeAttribute(GetName(wrapper));
 
                 Target newTarget = CreateTargetType(targetTypeName);
                 if (newTarget != null)
@@ -973,12 +966,6 @@ namespace NLog.Config
             }
 
             return false;
-        }
-
-        private static string GetConfigItemTypeAttribute(ILoggingConfigurationElement childElement, string sectionNameForRequiredValue = null)
-        {
-            var typeAttributeValue = sectionNameForRequiredValue != null ? childElement.GetRequiredValue("type", sectionNameForRequiredValue) : childElement.GetOptionalValue("type", null);
-            return StripOptionalNamespacePrefix(typeAttributeValue);
         }
 
         private bool ParseCompoundTarget(
@@ -1012,7 +999,7 @@ namespace NLog.Config
 
             if (IsTargetElement(name))
             {
-                string targetTypeName = GetConfigItemTypeAttribute(childElement, GetName(compound));
+                string targetTypeName = childElement.GetConfigItemTypeAttribute(GetName(compound));
 
                 Target newTarget = CreateTargetType(targetTypeName);
                 if (newTarget != null)
@@ -1036,9 +1023,7 @@ namespace NLog.Config
             return false;
         }
 
-
-        private void ConfigureObjectFromAttributes(object targetObject, ILoggingConfigurationElement element,
-            bool ignoreType)
+        private void ConfigureObjectFromAttributes(object targetObject, ILoggingConfigurationElement element, bool ignoreType)
         {
             foreach (var kvp in element.Values)
             {
@@ -1062,7 +1047,6 @@ namespace NLog.Config
                 }
             }
         }
-
 
         private void SetPropertyFromElement(object o, ILoggingConfigurationElement childElement, ILoggingConfigurationElement parentElement)
         {
@@ -1178,7 +1162,7 @@ namespace NLog.Config
                 return null;
 
             // Check if the 'type' attribute has been specified
-            string layoutTypeName = GetConfigItemTypeAttribute(element);
+            string layoutTypeName = element.GetConfigItemTypeAttribute();
             if (layoutTypeName == null)
                 return null;
 
@@ -1236,7 +1220,7 @@ namespace NLog.Config
 
         private Target WrapWithDefaultWrapper(Target target, ILoggingConfigurationElement defaultParameters)
         {
-            string wrapperTypeName = GetConfigItemTypeAttribute(defaultParameters, "targets");
+            string wrapperTypeName = defaultParameters.GetConfigItemTypeAttribute("targets");
             Target wrapperTargetInstance = CreateTargetType(wrapperTypeName);
             WrapperTargetBase wtb = wrapperTargetInstance as WrapperTargetBase;
             if (wtb == null)
@@ -1297,6 +1281,13 @@ namespace NLog.Config
             }
         }
 
+        private bool? ParseNullableBooleanValue(string propertyName, string value, bool defaultValue)
+        {
+            return StringHelpers.IsNullOrWhiteSpace(value)
+                ? (bool?)null
+                : ParseBooleanValue(propertyName, value, defaultValue);
+        }
+
         private bool MustThrowConfigException(NLogConfigurationException configException)
         {
             if (configException.MustBeRethrown())
@@ -1328,117 +1319,9 @@ namespace NLog.Config
                    || name.Equals("compound-target-ref", StringComparison.OrdinalIgnoreCase);
         }
 
-        /// <summary>
-        /// Remove the namespace (before :)
-        /// </summary>
-        /// <example>
-        /// x:a, will be a
-        /// </example>
-        /// <param name="attributeValue"></param>
-        /// <returns></returns>
-        private static string StripOptionalNamespacePrefix(string attributeValue)
-        {
-            if (attributeValue == null)
-            {
-                return null;
-            }
-
-            int p = attributeValue.IndexOf(':');
-            if (p < 0)
-            {
-                return attributeValue;
-            }
-
-            return attributeValue.Substring(p + 1);
-        }
-
         private static string GetName(Target target)
         {
             return string.IsNullOrEmpty(target.Name) ? target.GetType().Name : target.Name;
-        }
-    }
-
-    internal static class ILoggingConfigurationSectionExtensions
-    {
-        public static bool MatchesName(this ILoggingConfigurationElement section, string expectedName)
-        {
-            return string.Equals(section?.Name?.Trim(), expectedName, StringComparison.OrdinalIgnoreCase);
-        }
-
-        public static void AssertName(this ILoggingConfigurationElement section, params string[] allowedNames)
-        {
-            foreach (var en in allowedNames)
-            {
-                if (section.MatchesName(en))
-                    return;
-            }
-
-            throw new InvalidOperationException(
-                $"Assertion failed. Expected element name '{string.Join("|", allowedNames)}', actual: '{section?.Name}'.");
-        }
-
-        public static string GetRequiredValue(this ILoggingConfigurationElement element, string attributeName,
-            string section)
-        {
-            string value = element.GetOptionalValue(attributeName, null);
-            if (value == null)
-            {
-                throw new NLogConfigurationException($"Expected {attributeName} on {element.Name} in {section}");
-            }
-
-            if (StringHelpers.IsNullOrWhiteSpace(value))
-            {
-                throw new NLogConfigurationException(
-                    $"Expected non-empty {attributeName} on {element.Name} in {section}");
-            }
-
-            return value;
-        }
-
-        public static string GetOptionalValue(this ILoggingConfigurationElement section, string attributeName,
-            string defaultValue)
-        {
-            string value = section.Values
-                .Where(configItem => string.Equals(configItem.Key, attributeName, StringComparison.OrdinalIgnoreCase))
-                .Select(configItem => configItem.Value).FirstOrDefault();
-            if (value == null)
-            {
-                return defaultValue;
-            }
-
-            return value;
-        }
-
-        /// <summary>
-        /// Gets the optional boolean attribute value.
-        /// </summary>
-        /// <param name="section"></param>
-        /// <param name="attributeName">Name of the attribute.</param>
-        /// <param name="defaultValue">Default value to return if the attribute is not found or if there is a parse error</param>
-        /// <returns>Boolean attribute value or default.</returns>
-        public static bool GetOptionalBooleanValue(this ILoggingConfigurationElement section, string attributeName,
-            bool defaultValue)
-        {
-            string value = section.GetOptionalValue(attributeName, null);
-            if (string.IsNullOrEmpty(value))
-            {
-                return defaultValue;
-            }
-
-            try
-            {
-                return Convert.ToBoolean(value.Trim(), CultureInfo.InvariantCulture);
-            }
-            catch (Exception exception)
-            {
-                var configException = new NLogConfigurationException(exception, $"'{attributeName}' hasn't a valid boolean value '{value}'. {defaultValue} will be used");
-                if (configException.MustBeRethrown())
-                {
-                    throw configException;
-                }
-                InternalLogger.Error(exception, configException.Message);
-                return defaultValue;
-            }
         }
     }
 }
